// src/theme/types.ts
export interface EditorTheme {
  primary?: string;
  background?: string;
  text?: string;
  tabBarBackground?: string;
  sliderActive?: string;
  sliderInactive?: string;
  rulerBg?: string;
  rulerTickActive?: string;
  rulerTickInactive?: string;
  rulerPointer?: string;
  iconActive?: string;
  iconInactive?: string;
  toolButtonActiveBg?: string;
  toolButtonInactiveBg?: string;
}